Parte I terminada pero sin probar

+++ MyBatis_Introduction_VideoRental/src/main/java/edu/eci/cvds/samples/services/impl/ServiciosAlquilerItemsStub.java
@@ -39,7 +39,7 @@ public class ServiciosAlquilerItemsStub implements ServiciosAlquiler {
     }
 
     @Override
-    public int valorMultaRetrasoxDia(int itemId) {
+    public long valorMultaRetrasoxDia(int itemId) {
         return MULTA_DIARIA;
     }
